+++ frontend/src/types.ts
@@ -144,6 +144,7 @@ export const defaultProblemSet: ProblemSetSpec = {
 export type DocumentOptions = {
   fontSize: number;
   title: string | null;
+  subtitle: string | null;
   answerColumns: number;
   lang: 'Sv' | 'En';
   writeSolutions: 'None' | 'First' | 'All';
@@ -159,6 +160,7 @@ export type DocumentOptions = {
 export const defaultDocumentOptions: DocumentOptions = {
   fontSize: 10,
   title: null,
+  subtitle: null,
   answerColumns: 3,
   lang: 'Sv',
   writeSolutions: 'First',